LF-16425: mxc: vpu: hantro_v4l2: Calculate slice size more accurately
When enable the multi slice encoding, the hantro encoder calculate the
slice num according to the mb row.
But the v4l2 ctrl just set the mbs, not align to the mb rows.
so driver need to change the mbs to mb rows, and calculate the slice size
more accurately.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>