LF-16435 net: enetc: correct max RSC size
Currently, the max RSC size is set to ENETC_RS_MAX_BYTES(order), which
may cause the RX ring to use more than (MAX_SKB_FRAGS + 1) buffers to
receive a RSC frame. However, the size of struct skb_shared_info::frags
is MAX_SKB_FRAGS. So this will lead to the risk of out-of-bounds access
to the frags array. This is a potential risk and needs to be fixed.

Fixes: 897ff4adc024 ("LF-16411 net: enetc: add enetc_set_ringparam() to support changing rx-buf-len")
Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>