LF-16455: media: amphion: Disable flush mode to avoid performance loss
The flush mode of H264 and HEVC appends the padding code to indicate the
end of the frame, thereby telling the firmware that parsing can be
ended, with the aim of reducing latency.

But it led to firmware spend more time on parsing, except parsing
sequence header.

The parsing of each frame increase several milliseconds, this slightly
affects performance.

The root cause in the firmware is described below:

In normal mode, the vpu will try to parse more frames before decoding,
and in decoding, it will start parsing the next frame in parallel.

But in flush mode, the vpu will decode frame immediately after the frame
is parsed, there is no parallel.

And parsing one frame may take 1~2 milliseconds, this leads to
performance differences.

Normally we focus more on performance, so disable the flush mode by
default.

If users are more concerned about latency and can accept a slight
performance loss, then can manually turn on flush mode.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>