regulator: pca9450: resume i2c bus before restarting
Based on:
    https://github.com/atmark-techno/linux-5.10-at/commit/fba1f6a7f368d4c4f6d60566a096d693926951ef
    https://github.com/atmark-techno/linux-5.10-at/commit/73a9ddb1076e979faa0d39c42cfb70f6c0f04763

The i2c bus has likely been suspended and requires resuming, this is
done by keeping track of the i2c device and explicitely calling
pm_runtime_resume.

This is harmless if the device was not actually suspended, but the
function will then return an error so just continue after warning.

Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>