phy-fsl-samsung-hdmi: add pixel clock for 32.0/51.2/65.0/71.0MHz
Based on:
    https://github.com/atmark-techno/linux-5.10-at/commit/3383eee1bdc0183f9c8943fe2ff32435a73de9a6/
    https://github.com/atmark-techno/linux-5.10-at/commit/78ddc7670a12d03ecf5e50b85327312d1c07aba2/

Add pixel clock to use displays which do not support any standard HDMI
frequencies.

The frequencies are based on the previous closest value with a small
divider updated to provide a rough approximation, further work will be
done to further tune the values and/or automatically select something
approximate for other non-standard displays.

Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>