net: phy: dp83825: support TX data voltage tuning
DP83825 offers the possibility to tune the voltage of logical levels
of the MLT-3 encoded TX data. This is useful for adjusting the
waveform.

Unlike the dp83826, the dp83826 datasheet does not provide detailed
setting values ​​in the Register Description. Therefore, in the device
tree, specify the immediate value of the register.

Signed-off-by: Daisuke Mizobuchi <mizo@atmark-techno.com>